# Introduce Primitive.attr! to annotate 'inline'