https://bugs.ruby-lang.org/issues/15589

Based on https://github.com/ruby/ruby/pull/3226 which makes `Integer#zero?` safe for inlining, this PR introduces `Primitive.attr! 'inline'` to tell MJIT that builtin functions in an ISeq can be inlined. Note that [`Primitive`](https://bugs.ruby-lang.org/issues/16254) is private to Ruby core and it's not user-visible interface.

By just adding `Primitive.attr! 'inline'` on top of the `Integer#zero?` definition, `int.zero?` becomes almost as fast as `int == 0`.

```
$ benchmark-driver -v --rbenv 'before;before+JIT::before --jit;after+JIT::after --jit' --repeat-count=4 benchmark/mjit_int_zero_p.yml
before: ruby 2.8.0dev (2020-06-20T21:55:30Z master cc14319b71) [x86_64-linux]
before+JIT: ruby 2.8.0dev (2020-06-20T21:55:30Z master cc14319b71) +JIT [x86_64-linux]
after+JIT: ruby 2.8.0dev (2020-06-20T22:02:33Z builtin-attr 9434fe34a8) +JIT [x86_64-linux]
Calculating -------------------------------------
                         before  before+JIT   after+JIT
             0.zero?    42.950M     54.840M     71.566M i/s -     40.000M times in 0.931325s 0.729394s 0.558922s
             1.zero?    43.278M     55.048M     72.340M i/s -     40.000M times in 0.924253s 0.726642s 0.552948s
              0 == 0    63.616M     78.808M     78.406M i/s -     40.000M times in 0.628776s 0.507566s 0.510166s
              1 == 0    63.824M     79.360M     79.173M i/s -     40.000M times in 0.626721s 0.504034s 0.505224s

Comparison:
                          0.zero?
           after+JIT:  71566276.6 i/s
          before+JIT:  54840015.0 i/s - 1.31x  slower
              before:  42949584.0 i/s - 1.67x  slower

                          1.zero?
           after+JIT:  72339552.3 i/s
          before+JIT:  55047758.6 i/s - 1.31x  slower
              before:  43278202.5 i/s - 1.67x  slower

                           0 == 0
          before+JIT:  78807508.3 i/s
           after+JIT:  78405894.2 i/s - 1.01x  slower
              before:  63615623.3 i/s - 1.24x  slower

                           1 == 0
          before+JIT:  79359754.1 i/s
           after+JIT:  79172860.9 i/s - 1.00x  slower
              before:  63824291.7 i/s - 1.24x  slower
```

In this PR, there's no mechanism that verifies an ISeq with `Primitive.attr! 'inline'` is really safe for inlining. 

But we already know what to be checked https://bugs.ruby-lang.org/issues/16956, and we prove that kind of thing can be checked like `vm_check_canary` for the `leaf` attribute. So it'll be implemented once this gets merged, before starting to use the annotation in many places.